Extract JS code outside Activity
